experiments/pyproject.toml
[project]
name = "notebooks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "pydantic-ai>=0.0.52",
    "pydantic-ai-slim[openai]>=0.0.52",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
